mnsound: merge to sound
