+++ src/game/setup.c
@@ -43,7 +43,6 @@
 #include "game/items.h"
 #include "game/lara.h"
 #include "game/lot.h"
-#include "game/mnsound.h"
 #include "game/objects/boat.h"
 #include "game/objects/bridge.h"
 #include "game/objects/cabin.h"
@@ -60,6 +59,7 @@
 #include "game/objects/trapdoor.h"
 #include "game/overlay.h"
 #include "game/savegame.h"
+#include "game/sound.h"
 #include "game/text.h"
 #include "game/traps/damocles_sword.h"
 #include "game/traps/dart.h"
